services/trainer/pyproject.toml
[tool.poetry]
name = "trainer"
version = "0.1.0"
description = "Offline fine-tuning jobs for PB2A"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
evals = { path = "../../packages/evals", develop = true }
memory = { path = "../../packages/memory", develop = true }
transformers = "^4.35.0"
peft = "^0.6.0"
trl = "^0.7.4"
torch = "^2.1.0"
datasets = "^2.14.0"
PyYAML = "^6.0"
requests = "^2.31.0"
tqdm = "^4.66.0"
accelerate = "^0.26.0"
scipy = "^1.11.0"
sentencepiece = "^0.1.99"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
